requirements.txt
fastapi[standard]
pydantic-settings

# security
argon2-cffi
python-jose[cryptography] # for JWT handling

# Testing
pytest
httpx

# MongoDB (async driver)
motor
pymongo

# Redis (async client)
redis